skills/global/meta_skills_4erctua4/pyproject.toml
[project]
name = "antigravity-claw"
version = "1.0.0"
description = "Antigravity Claw — make any webpage defy gravity, powered by Rust + Python"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "websockets>=12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
